Made about page home page
