+++ src/app/home.css
@@ -4,7 +4,7 @@ nav {
 }
 
 /* General */
-.about-page {
+.home-page {
   font-family: 'Poppins', sans-serif;
   color: #333;
   background: #fff;
